==> AI/recsys/03_naver.ipynb <==
from selenium import webdriver
from selenium.webdriver.common.by import By
import pyautogui
from bs4 import BeautifulSoup
import requests

import pandas as pd
import time

import pickle
from tqdm import tqdm

### 상품 URL 추출

driver = webdriver.Chrome()
url = 'https://shopping.naver.com/beauty/category?menu=10003291'
driver.get(url)

array = []

SCROLL_PAUSE_SEC = 3
last_height = driver.execute_script("return document.body.scrollHeight")

while True:
    driver.execute_script("window.scrollTo(0, document.body.scrollHeight);")
    time.sleep(SCROLL_PAUSE_SEC)
    new_height = driver.execute_script("return document.body.scrollHeight")
    
    SCROLL_PAUSE_SEC = 5
    tmp = driver.find_elements(By.CSS_SELECTOR, "._2KkiZqJBZF > a")
#     array.append(tmp[0].get_attribute('href'))
#     print(len(array))
    
    for i in range(len(tmp)):
        array.append(tmp[i].get_attribute('href'))

    if new_height == last_height:
        try:
            driver.find_element_by_css_selector(".mye4qd").click()
        except:
            break
    last_height = new_height
    
    if len(array) >= 5000:
        break

len(array)

with open('naver_url.pickle', 'wb') as f:
    pickle.dump(array, f)

with open('naver_url.pickle', 'rb') as f:
    array = pickle.load(f)

array

### 각 상품 정보 추출

#### 1번 상품

# 1번 상품
url = f"{array[0]}"
response = requests.get(url)
dom = BeautifulSoup(response.text, 'html.parser')

# 상품명
name = dom.select("._3oDjSvLwq9")[0].text
# 가격
price = dom.find("span",{"class":"_1LY7DqCnwR"}).text

# 데이터 프레임 컬럼
find_columns = dom.select("._1Hbih69XFT")[0]
columns = find_columns.find_all("th",{"class":"_15qeGNn6Dt"})
df_columns = ["상품명", "가격"]
for i in range(len(columns)):
    df_columns.append(columns[i].text)

# 데이터 프레임 로우
find_rows = dom.select("._1Hbih69XFT")[0]
rows = find_rows.find_all("td",{"class":"jvlKiI0U_y"})
df_rows = [name, price]
for i in range(len(rows)):
    df_rows.append(rows[i].text)

# 데이터프레임 형성
df = pd.DataFrame(columns = df_columns)
df.loc[0] = df_rows

#### 2번부터

for i in tqdm(array[1:]):
    url = f"{i}"
    response = requests.get(url)
    dom = BeautifulSoup(response.text, 'html.parser')
    
    # 상품명
    name = dom.select("._3oDjSvLwq9")[0].text
    # 가격
    price = dom.find("span",{"class":"_1LY7DqCnwR"}).text
    
    # 데이터 프레임 컬럼
    find_columns = dom.select("._1Hbih69XFT")[0]
    columns = find_columns.find_all("th",{"class":"_15qeGNn6Dt"})
    df_columns = ["상품명", "가격"]
    for i in range(len(columns)):
        df_columns.append(columns[i].text)

    # 데이터 프레임 로우
    find_rows = dom.select("._1Hbih69XFT")[0]
    rows = find_rows.find_all("td",{"class":"jvlKiI0U_y"})
    df_rows = [name, price]
    for i in range(len(rows)):
        df_rows.append(rows[i].text)

    # 데이터프레임 형성
    n_df = pd.DataFrame(columns = df_columns)
    n_df.loc[0] = df_rows
    
    # 데이터프레임 병합
    # 기존에 만들어진 df, n_df = n_df
    df = pd.concat([df, n_df], axis=0)

df.isnull().sum()

### 데이터프레임 저장

df.to_csv("naver.csv")

pd.read_csv('naver.csv').head()
