==> AI/recsys/04_parsing.ipynb <==
import pandas as pd

train = pd.read_csv("naver.csv")
train.drop(['Unnamed: 0'], axis=1, inplace=True)

train.info()

train.head()

train_nan = list(train.isna().sum())
num_nan = list(filter(lambda x:x >= 1, train_nan))

index_nan = list(filter(lambda e:train_nan[e] >= 1, range(len(train_nan))))

nan_df = pd.DataFrame({'index_in_train' : index_nan, 'column_name' : train.iloc[:, index_nan].columns, 'NAN_number' : num_nan})
display(nan_df)

train

tmp = train['세부제품특징'].dropna()
result = []

for i in tmp:
    result += (i.split(","))
    
result2 = []
for i in result:
    result2.append(i.replace(" ", ""))
    
result2 = set(result2)
result2 = list(result2)
len(result2)

result2.sort()
print(len(tmp))
print(result2)

tmp = train['주요제품특징'].dropna()
result = []

for i in tmp:
    result += (i.split(","))
    
result3 = []
for i in result:
    result3.append(i.replace(" ", ""))
    
result3 = set(result3)
result3 = list(result3)
len(result3)

result3.sort()
print(len(tmp))
print(result3)

asd = pd.DataFrame()



for i in result2:
    x = []
    for n in train['주요제품특징'].dropna():
        if i in n:
            x.append(1)
        else:
            x.append(0)
    asd = pd.concat([asd, pd.DataFrame({ i : x })], axis = 1)
    
asd

data2 = train[['상품명', '가격', '세부제품특징']]

data2 = data2.dropna(subset=['세부제품특징'], axis = 0).reset_index(drop=True)

asd = asd.reset_index(drop=True)

concat_test = pd.concat([data2, asd], axis = 1)

display(concat_test.head())

concat_test2 = concat_test.dropna()

concat_test2['가격'] = concat_test2['가격'].str.replace(',', '')
concat_test2['가격'] = concat_test2['가격'].astype(int)

concat_test2.drop(['세부제품특징'], axis=1, inplace=True)

target_feat = '상품명'

x = concat_test2.drop(target_feat, axis = 1) # 이게 feature가 되고
y = concat_test2.loc[:, target_feat] # 이게 target이 됨

y

from sklearn.model_selection import train_test_split
x_train, x_valid, y_train, y_valid = train_test_split(x, y, test_size=0.3, random_state=2022)

from sklearn.tree import DecisionTreeClassifier

dt = DecisionTreeClassifier()
dt.fit(x_train,y_train)

pred = dt.predict(x_valid)

pred[0]

y_valid[0:10]
