==> AI/recsys/05_modeling.ipynb <==
import numpy as np
import pandas as pd

# Preprocessing
from sklearn.model_selection import train_test_split

# Model
from sklearn.tree import DecisionTreeClassifier

df = pd.read_csv("naver.csv")
df.drop(['Unnamed: 0'], axis=1, inplace=True)

train_nan = list(df.isna().sum())
num_nan = list(filter(lambda x:x >= 1, train_nan))

index_nan = list(filter(lambda e:train_nan[e] >= 1, range(len(train_nan))))

nan_df = pd.DataFrame({'index_in_train' : index_nan, 'column_name' : df.iloc[:, index_nan].columns, 'NAN_number' : num_nan})
display(nan_df)

df.info()

# Column Select
train_columns = ["상품명", "가격", "세부제품특징", 
                 "피부타입", "사용부위", "사용시간", "제품형태"]
df = df.loc[:, train_columns]

### NaN Preprocessing

# 유니크값 확인 함수
def unique(column):
    result = []
    result2 = []
    tmp = column.dropna()
    
    for i in tmp:
        result += i.split(",")
        
    for j in result:
        result2.append(j.replace(" ", ""))
        
    result2 = set(result2)
    result2 = list(result2)
    result2.sort()
    
    return result2

# 가변수화 함수
def dummy(column):
    tmp = pd.DataFrame()
    for i in unique(column):
        x = []
        for n in column.dropna():
            if i in n:
                x.append(1)
            else:
                x.append(0)
        tmp = pd.concat([tmp, pd.DataFrame({ i : x })], axis = 1)
    return tmp

# 결측치 처리
df.isna().sum()

# 세부제품특징

# parsing
unique(df['세부제품특징'])

# 피부타입 : 모든피부용
df['피부타입'] = df['피부타입'].fillna('모든피부용')

# 사용부위

# parsing
unique(df['사용부위'])

# 사용시간

df['사용시간'] = df['사용시간'].fillna('낮, 밤')

# 제품형태

# parsing
unique(df['제품형태'])

df.isnull().sum()

# drop na
df = df.dropna()
df = df.reset_index(drop=True)
df

### Column Preprocess

# 가격
df['가격'] = df['가격'].str.replace(',', '')
df['가격'] = df['가격'].astype(int)

# 세부제품특징
tmp = dummy(df['세부제품특징'])
df = pd.concat([df, tmp], axis = 1)
df.drop(['세부제품특징'], axis=1, inplace=True)

# 피부타입
tmp = dummy(df['피부타입'])
df = pd.concat([df, tmp], axis = 1)
df.drop(['피부타입'], axis=1, inplace=True)

# 사용부위
tmp = dummy(df['사용부위'])
df = pd.concat([df, tmp], axis = 1)
df.drop(['사용부위'], axis=1, inplace=True)

# 사용시간
df['사용시간'] = df['사용시간'].str.replace('낮, 밤', '0')
df['사용시간'] = df['사용시간'].str.replace('밤, 낮', '0')
df['사용시간'] = df['사용시간'].str.replace('낮', '1')
df['사용시간'] = df['사용시간'].str.replace('밤', '2')
df['사용시간'] = df['사용시간'].astype('int')

# 제품형태
tmp = dummy(df['제품형태'])
df = pd.concat([df, tmp], axis = 1)
df.drop(['제품형태'], axis=1, inplace=True)

df.columns

### Modeling

# x, y split
target = '상품명'
x = df.drop(target, axis = 1)
y = df.loc[:, target]

# train, valid split
x_train, x_valid, y_train, y_valid = train_test_split(x, y, test_size=0.3, random_state=2022)

dt = DecisionTreeClassifier()
dt.fit(x_train, y_train)

import joblib

joblib.dump(dt, 'recsys_dt.pkl')

pred = dt.predict(x_valid)

pred[0:10]

y_valid[0:10]

### 유저 데이터 넣기

user = [0, '노폐물 제거', '모든피부용', '페이스용', '낮, 밤', '스크럽형']

test_df = pd.DataFrame(columns=df.columns)
test_df

test_df.loc[0, :] = user
test_df

# 가격
try:
    test_df['가격'] = test_df['가격'].str.replace(',', '')
    test_df['가격'] = test_df['가격'].astype(int)
except:
    test_df['가격'] = test_df['가격'].astype(int)

# 세부제품특징
tmp = dummy(test_df['세부제품특징'])
test_df = pd.concat([test_df, tmp], axis = 1)
test_df.drop(['세부제품특징'], axis=1, inplace=True)

# 피부타입
tmp = dummy(test_df['피부타입'])
test_df = pd.concat([test_df, tmp], axis = 1)
test_df.drop(['피부타입'], axis=1, inplace=True)

# 사용부위
tmp = dummy(test_df['사용부위'])
test_df = pd.concat([test_df, tmp], axis = 1)
test_df.drop(['사용부위'], axis=1, inplace=True)

# 사용시간
test_df['사용시간'] = test_df['사용시간'].str.replace('낮, 밤', '0')
test_df['사용시간'] = test_df['사용시간'].str.replace('밤, 낮', '0')
test_df['사용시간'] = test_df['사용시간'].str.replace('낮', '1')
test_df['사용시간'] = test_df['사용시간'].str.replace('밤', '2')
test_df['사용시간'] = test_df['사용시간'].astype('int')

# 제품형태
tmp = dummy(test_df['제품형태'])
test_df = pd.concat([test_df, tmp], axis = 1)
test_df.drop(['제품형태'], axis=1, inplace=True)

test_df
