==> AI/recsys/06_user2test.ipynb <==
import numpy as np
import pandas as pd

# Preprocessing
from sklearn.model_selection import train_test_split

# Model
from sklearn.tree import DecisionTreeClassifier

import joblib, pickle

model = joblib.load('recsys_dt.pkl')
with open('df_columns', 'rb') as f:
    columns = pickle.load(f)

df_url = pd.read_csv("naver_url.csv")
df_url = df_url[['상품명', 'url']]
df_url.head()

# df 생성
test_df = pd.DataFrame(columns=columns)
test_df = test_df.iloc[:, 1:]
test_df

test_df.loc[0, :] = [0] * (len(columns) - 1)
test_df

# 유저 데이터
user = ['노폐물제거', '모든피부용', '페이스용', '스크럽형']

test_df.columns

for u in user:
    tmp = u
    test_df[tmp] += 1
    
test_df

pred = model.predict(test_df)

pred[0]

df_url['url'][df_url['상품명'] == pred[0]]
